ts 변환 로직 추가
