+++ src/convert_v7/7version.ts
@@ -1,5 +1,7 @@
 interface V7 {
-  (objColumn: any, str: string): void;
+  convertTreeData: (data7: any) => any;
+  convertDateFormat: (obj: any) => void;
+  convertAcceptKeys: (objColumn: any, str: any) => void;
   (obj: any): void;
   (data7: any[]): void;
 }
@@ -8,20 +10,9 @@ interface IBSHEETCONVERT {
   v7: V7
 }
 
-interface IB_OBJ {
+export interface IB_OBJ {
   [index: string]: any
 }
 
 export let IBSheetConvert: IBSHEETCONVERT;
 
-function clone(obj: IB_OBJ | null): object {
-  if (obj === null || typeof obj !== 'object') return obj as any;
-  const copy = obj.constructor();
-
-  for (let attr in obj) {
-    if (obj.hasOwnProperty(attr)) {
-      copy[attr] = clone(obj[attr]);
-    }
-  }
-  return copy;
-}
